% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.cell.sim.R
\name{my.cell.sim}
\alias{my.cell.sim}
\title{Cell transformation simulation}
\usage{
my.cell.sim(n = 5, k = 5, n.mesh = 32, n.step = 50, scale.shift = 0.1,
  scale.rotation1 = 0.1, scale.rotation2 = 0.1, scale.shear = 0.1,
  file.name = "hoge")
}
\arguments{
\item{n}{depth of sphere harmonic functions}

\item{k}{simple sphere radius}

\item{n.mesh}{fineness of mesh}

\item{n.step}{number of time steps}

\item{scale.shift}{scale of shift}

\item{scale.rotation1}{scale of rotation 1}

\item{scale.rotation2}{scale of rotation 2}

\item{scale.shear}{scale of shear}

\item{file.name}{output file name}
}
\description{
Simulate cell transformation and write series of obj files
}
\examples{
library(devtools)
install_github("ryamada22/Ronlyryamada")
library(Ronlyryamada)
library(rgl)
library(RFOC)
# parameters
n <- 5 # depth of sphere harmonic functions
k <- 5 # simple sphere radius
n.mesh <- 32 # fineness of mesh
n.step <- 50 # number of time steps
scale.shift <- 0.1 # shift
scale.rotation1 <- 0.1 # rotation 1
scale.rotation2 <- 0.1 # rotation 2
scale.shear <- 0.1 # shear
file.name <- "hoge"
my.cell.sim()
}

