% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.circ.measure.R
\name{my.circ.measure}
\alias{my.circ.measure}
\title{2D volume and circumference}
\usage{
my.circ.measure(x, r)
}
\arguments{
\item{x}{is a 0/1 matrix.}

\item{r}{is a positive real value of width to measure circumference.}
}
\description{
These functions for 2D 0/1 image matrices
}
\examples{
x <- matrix(1:16,4,4)
v <- c(2,3)

my.parallel(x,v)
v <- c(-2,-3)
my.parallel(x,v)
v <- c(2,-3)
my.parallel(x,v)
tmp.x.list <- array(0,c(dim(xy.mat),length(out[,1])))
for(i in 1:length(out[,1])){
	tmp.v <- out[i,]
	tmp.x.list[,,i] <- my.parallel(xy.mat,tmp.v)
}
tt <- apply(tmp.x.list,c(1,2),sum) - length(out[,1])*xy.mat
image(tt)
image(sign(tt))
image(tt !=0)
n <- 2^8
x <- y <- 1:n
xy <- as.matrix(expand.grid(x,y))
xy.val <- rep(0,length(xy[,1]))
n.circle <- 100
ctr <- matrix(sample(50:200,replace=TRUE,n.circle*2),ncol=2)
r <- sample(1:40,replace=TRUE,n.circle)

for(i in 1:n.circle){
	tmp.x <- xy[,1] - ctr[i,1]
	tmp.y <- xy[,2] - ctr[i,2]
	s <- which(tmp.x^2+tmp.y^2 < r[i]^2)
	xy.val[s] <- 1
}
xy.mat <- matrix(xy.val,ncol=n)
image(xy.mat)
rs <- seq(from=0,to=20,length=21)
rs <- rs[-1]
tmp.out <- list()
for(i in 1:length(rs)){
	tmp.out[[i]] <- my.circ.measure(xy.mat,rs[i])
}
circs <- rep(0,length(rs))
for(i in 1:length(tmp.out)){
	circs[i] <- tmp.out[[i]]$circ
}
plot(rs,circs)
}
\keyword{circumference}
\keyword{shape}

