% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.DECconformal.R
\name{my.curvature.cot}
\alias{my.curvature.cot}
\title{局所平均曲率を計算する関数
入力は頂点の四元数座標ベクトルと３行で表された三角形頂点ID行列
返り値はnorm.v,norm.face=ret.face,dirの３つで
norm.vは各頂点の法線ベクトルに相当する四元数の行列で、その虚部の絶対値が平均曲率の大きさ
norm.faceは各三角形のそれで、虚部の絶対値が平均曲率の大きさ
dirは各三角形の平均曲率が曲面の外向きか内向きか(今、扱っている曲面はすべて球面様の閉曲面なので、外向き・内向きはその意味での内外)を示すc(-1,0,1)}
\usage{
my.curvature.cot(vertices, faces.v)
}
\description{
局所平均曲率を計算する関数
入力は頂点の四元数座標ベクトルと３行で表された三角形頂点ID行列
返り値はnorm.v,norm.face=ret.face,dirの３つで
norm.vは各頂点の法線ベクトルに相当する四元数の行列で、その虚部の絶対値が平均曲率の大きさ
norm.faceは各三角形のそれで、虚部の絶対値が平均曲率の大きさ
dirは各三角形の平均曲率が曲面の外向きか内向きか(今、扱っている曲面はすべて球面様の閉曲面なので、外向き・内向きはその意味での内外)を示すc(-1,0,1)
}

