% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.plot.flat.circle.R
\name{my.fit.sphere}
\alias{my.fit.sphere}
\title{place on sphere with radius R}
\usage{
my.fit.sphere(x, R = 1)
}
\arguments{
\item{x, }{a matrix; num. of column = dimension}

\item{R}{is a real value of radius}
}
\description{
Fit to sphere
}
\examples{
x <- matrix(rnorm(100*3),ncol=3)
x <- my.fit.sphere(x)
}

