% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.HadamardWalk.R
\name{my.HadamardWalk}
\alias{my.HadamardWalk}
\title{Hadamard walk}
\usage{
my.HadamardWalk(U, Xinit, n)
}
\arguments{
\item{U}{is an Hermitian 2x2 matrix.}

\item{Xinit}{is an complex vector with length 2.}

\item{n}{is an number of steps.}
}
\description{
Hadamard walk simulation
}
\examples{
U <- 1/sqrt(2) * matrix(c(1,1,1,-1),byrow=TRUE,2,2) + 1i*0
Xinit <- 1/sqrt(2) * c(1,1i)
hout <- my.HadamardWalk(U,Xinit,50)
image(hout[[3]])
persp(hout[[3]])
}
\keyword{Quantum}
\keyword{walk}

