% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.halfedge.R
\name{my.halfedge}
\alias{my.halfedge}
\title{quaternion matrix}
\usage{
my.halfedge(xyz, faces.v)
}
\description{
quaternion matrix
}
\examples{
library(Matrix)
example(my.catmull.clark.tri)
mesh <- my.mesh(faces.v)
xyz <- t(as.matrix(vertices)[2:4,])
my.mesh.tri.plot(vertices,faces.v)
area.norm <- my.area.norm(mesh,xyz)
cot <- my.halfedge.cot(mesh,xyz)
dec <- my.dec(mesh,xyz)
n.iter <- 100
current.xyz <- xyz
for(i in 1:n.iter){
	new.out <- my.update.heatflow(mesh,current.xyz,step=0.05)
	my.mesh.tri.plot.xyz(new.out$xyz,faces.v)
	current.xyz <- new.out$xyz
}
}

