% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfManifoldHyperCube.R
\name{my.haplotypes}
\alias{my.haplotypes}
\title{e-flat and m-flat parameterizations for {0,1}^d}
\usage{
my.haplotypes(d)
}
\description{
e-flat and m-flat
}
\examples{
library(MCMCpack)
freq <- c(rdirichlet(1,rep(1,n)))
freq
thetapsi <- my.freq2thetapsi(freq)
psi <- thetapsi[1]
theta <- thetapsi[-1]
thetapsi
eta <- my.freq2eta(freq)
eta
my.eta2freq(eta)
my.eta2thetapsi(eta)
my.theta2freq(theta)
my.theta2eta(theta)

freq <- rep(1/8,8)
freq
thetapsi <- my.freq2thetapsi(freq)
psi <- thetapsi[1]
theta <- thetapsi[-1]
thetapsi
eta <- my.freq2eta(freq)
eta
my.eta2freq(eta)
my.eta2thetapsi(eta)
my.theta2freq(theta)
my.theta2eta(theta)

dd <- 0
freq <- c(1/2-dd*3,dd,dd,dd,dd,dd,dd,1/2-dd*3)
freq
thetapsi <- my.freq2thetapsi(freq)
psi <- thetapsi[1]
theta <- thetapsi[-1]
thetapsi
eta <- my.freq2eta(freq)
eta
my.eta2freq(eta)
my.eta2thetapsi(eta)
my.theta2freq(theta)
my.theta2eta(theta)
}

