% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.DECconformal.R
\name{my.inv.pow.2}
\alias{my.inv.pow.2}
\title{Ax = lambda b の固有ベクトルを逆冪乗法で近似する
初期ベクトル bは(1,1,...)がよいことが知られているので
それをデフォルト値としてある}
\usage{
my.inv.pow.2(A, n.iter = 3, b = rep(1, ncol(A)), log = FALSE)
}
\description{
Ax = lambda b の固有ベクトルを逆冪乗法で近似する
初期ベクトル bは(1,1,...)がよいことが知られているので
それをデフォルト値としてある
}

