% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.labeling.R
\name{my.labeling}
\alias{my.labeling}
\title{Arbitrary dimensional voxel data labeling
This function labels continuous segments.}
\usage{
my.labeling(data.)
}
\arguments{
\item{data.}{is a matrix with ncol-dimensions}
}
\description{
Arbitrary dimensional voxel data labeling
This function labels continuous segments.
}
\examples{
library(igraph)
df <- 4
n <- 20
data. <- array(0,rep(n,df))
addr <- t(which(data.>=0,arr.ind=TRUE))
n.sp <- 20
for(i in 1:n.sp){
	tmp <- runif(df) * n
	tmp2 <- runif(1) * n/5
	data.[apply((addr - tmp)^2,2,sum) < tmp2] <- 1
}
clu <- my.labeling(data.)
hist(clu[[3]],ylim=c(0,100))
cluster.n <- clu[[1]][[3]]
vol.slice <- list()
dm <- dim(data.)
for(i in 1:cluster.n){
	tmp.arr <- array(as.numeric(clu[[3]]==i),dm)
	vol.slice[[i]] <- my.slice.vol(tmp.arr,df)
}
sapply(vol.slice,sum)
clu[[1]][[2]]
dm <- c(2^8,2^8,2^3,2^4)
ar <- array(1:prod(dm),dm)
d <- sample(1:length(dm),1)
j <- sample(1:dm[d],1)
out1 <- my.slice(c(ar),dm,d,j)
out2 <- my.slice.2(c(ar),dm,d,j)
plot(out2[[1]])
}

