% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.sp.harmonics.R
\name{my.Legendre}
\alias{my.Legendre}
\title{Spherical harmonics}
\usage{
my.Legendre(k)
}
\arguments{
\item{x, }{a matrix; num. of column = dimension}

\item{R}{is a real value of radius}
}
\description{
Spherical harmonics
}
\examples{
require(rgl)
require(RFOC)
require(Ronlyryamada)
k <- 20
Pms <- my.Legendre(k)
x <- seq(from=-1,to=1,length=100)
ys <- lapply(Pms,my.poly.calc,x)
plot(x,ys[[1]],type="l",ylim=c(-1,1))
for(i in 2:length(ys)){
	points(x,ys[[i]],type="l",col=i)
}
n <- 6
A <- matrix(0,n,n)
ret <- matrix(list(),n,n)
for(i in 1:n){
	for(j in 1:i){
		A. <- A
		A.[i,j] <- 1
		ret[i,j] <- list(my.spherical.harm(A.))
	}
}
tmp <- ret[4,1][[1]]
out <- rbind(tmp$X,rep(min(tmp$X),3),rep(max(tmp$X),3))
plot3d(out,type="l",col=rainbow(128))
col1 <- col2 <- rep(0,length(tmp$r))
col1[which(tmp$r>=0)] <- tmp$r[which(tmp$r>=0)]
col2[which(tmp$r<0)] <- -tmp$r[which(tmp$r<0)]
col1 <- col1/max(col1)
col2 <- col2/max(col2)
plot3d(tmp$X.,col=rgb(1-col1,1-col2,0),size=10)
n <- 6
A <- matrix(0,n,n)
ret <- matrix(list(),n,n)
for(i in 1:n){
	for(j in 1:i){
		A. <- A
		A.[i,j] <- 1
		ret[i,j] <- list(my.spherical.harm(A.))
	}
}
# ”¼Œa‚»‚Ì‚Ü‚Ü
tmp <- ret[4,1][[1]]
out <- rbind(tmp$X,rep(min(tmp$X),3),rep(max(tmp$X),3))
plot3d(out,type="l",col=rainbow(128))
# ”¼Œa‚Ì‚Qæ‚Å
out.. <- rbind(tmp$X*tmp$r,rep(min(tmp$X*tmp$r),3),rep(max(tmp$X*tmp$r),3))
plot3d(out..,type="l",col=rainbow(128))
# ”¼Œa‚ðF‚Å
col1 <- col2 <- rep(0,length(tmp$r))
col1[which(tmp$r>=0)] <- tmp$r[which(tmp$r>=0)]
col2[which(tmp$r<0)] <- -tmp$r[which(tmp$r<0)]
col1 <- col1/max(col1)
col2 <- col2/max(col2)
plot3d(tmp$X.,col=rgb(1-col1,1-col2,0),size=10)
n <- 6
A <- matrix(0,n,n)
ret <- matrix(list(),n,n)
for(i in 1:n){
	for(j in 1:i){
		A. <- A
		A.[1,1] <- 5 # ³‰~‚Ì•ª
		A.[i,j] <- 1
		ret[i,j] <- list(my.spherical.harm(A.))
	}
}
tmp <- ret[4,1][[1]]
out <- rbind(tmp$X,rep(min(tmp$X),3),rep(max(tmp$X),3))
plot3d(out,type="l",col=rainbow(128))
n <- 6
A. <- matrix(runif(n^2),n,n)
A.[1,1] <- 5
xxx <- my.spherical.harm.mesh(A.,n=32)
plot3d(xxx$v)
segments3d(xxx$v[c(t(xxx$edge)),])
}

