% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.DECconformal.R
\name{my.make.E.v}
\alias{my.make.E.v}
\title{E行列作成関数
ただし、四元数行列なので、虚実４成分それぞれをリストの要素とし
個々のリストも行列ではなく、ベクトル化した形で返す
疎行列のリスト
三角形の面積を考慮して、面のrho値に応じて、rhoが作る曲面上の
ベクトル場を作り
その定常解の一つとして固有ベクトルを取る}
\usage{
my.make.E.v(vertices, faces.v, rho)
}
\description{
E行列作成関数
ただし、四元数行列なので、虚実４成分それぞれをリストの要素とし
個々のリストも行列ではなく、ベクトル化した形で返す
疎行列のリスト
三角形の面積を考慮して、面のrho値に応じて、rhoが作る曲面上の
ベクトル場を作り
その定常解の一つとして固有ベクトルを取る
}

