% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.DECconformal.R
\name{my.make.L}
\alias{my.make.L}
\title{ラプラシアン行列の算出
すべての点について、その点を含む三角形について、三角形の内角で定まる値を加算する}
\usage{
my.make.L(vertices, faces.v)
}
\description{
ラプラシアン行列の算出
すべての点について、その点を含む三角形について、三角形の内角で定まる値を加算する
}

