% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.DECconformal.R
\name{my.make.quatList}
\alias{my.make.quatList}
\title{ラプラシアン行列Lは実行列だが、その四元数版の実行列化したもの(行・列がそれぞれ４倍)を作る}
\usage{
my.make.quatList(L)
}
\description{
ラプラシアン行列Lは実行列だが、その四元数版の実行列化したもの(行・列がそれぞれ４倍)を作る
}

