% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.peri.tri.R
\name{my.peri.tri}
\alias{my.peri.tri}
\title{3D object surface triangles}
\usage{
my.peri.tri(v, h = max(v), smooth = 0)
}
\arguments{
\item{v}{is a 3d-array..}

\item{h}{is a real value to draw contour line.}
}
\description{
These function returns surface trinangles of 3D object
}
\examples{
library(misc3d)
k <- 4
out <- my.3d.obj(3,k)
obj <- out$xyz[which(out$xyz.val==1),]
library(rgl)
plot3d(obj)

xyz.arr <- array(out$xyz.val,c(2^k,2^k,2^k))
tris <- my.peri.tri(xyz.arr)
drawScene.rgl(tris)
}
\keyword{mesh}
\keyword{shape}
\keyword{triangulation}

