% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.plot.flat.circle.R
\name{my.plot.hemispheres}
\alias{my.plot.hemispheres}
\title{Plot two hemi-spheres}
\usage{
my.plot.hemispheres(x)
}
\arguments{
\item{x}{2-column matrix with angles}
}
\description{
plot two hemi-spheres
}
\examples{
require(RFOC)
x <- matrix(rnorm(1000*3),ncol=3)
x <- my.fit.sphere(x)
x.sp <- TOSPHERE(x[,1],x[,2],x[,3])
plot(x.sp[[1]],x.sp[[2]])
x.sp.2 <- list(x.sp[[1]]/360*2*pi,x.sp[[2]]/360*2*pi)
my.plot.hemispheres(x.sp.2)
}

