% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.spin.transformation.R
\name{my.spin.transformation}
\alias{my.spin.transformation}
\title{Spin transformation}
\usage{
my.spin.transformation(n = 5, k = runif(1), n.mesh = 32, n.step = 10,
  scale.shift = 0.1, scale.rotation1 = 0.1, scale.rotation2 = 0.1,
  scale.shear = 0.1, file.name = "hoge")
}
\arguments{
\item{n}{depth of sphere harmonic functions}

\item{k}{simple sphere radius}

\item{n.mesh}{fineness of mesh}

\item{n.step}{number of time steps}

\item{scale.shift}{scale of shift}

\item{scale.rotation1}{scale of rotation 1}

\item{scale.rotation2}{scale of rotation 2}

\item{scale.shear}{scale of shear}

\item{file.name}{output file name}
}
\description{
Spin transformation
}
\examples{
library(rgl)
library(RFOC)
library(onion)
out3 <- my.spin.transformation(n.mesh=64,scale.shift=3,scale.shear=0.2)
plot3d(out3$new.xyz,xlim=range(out3$new.xyz),ylim=range(out3$new.xyz),zlim=range(out3$new.xyz))
segments3d(out3$new.xyz[c(t(out3$mesh$edge)),])
}

