% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.Unitary2x2.R
\name{my.Unitary2x2}
\alias{my.Unitary2x2}
\title{2x2 Unitary Matrix with determinant +/- 1}
\usage{
my.Unitary2x2(theta, phi1, phi2)
}
\arguments{
\item{theta, }{a real value.}

\item{phi1, }{a real value.}

\item{phi2, }{a real value.}
}
\description{
2x2 Unitary Matrix
}
\examples{
theta <- pi/6;phi1 <- pi/4;phi2 <- pi/3
U <- my.Unitary2x2(theta,phi1,phi2)
U
}
\keyword{Unitary}
\keyword{matrix}

