% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.DECconformal.R
\name{my.vector.access}
\alias{my.vector.access}
\title{Utility 関数
特に疎行列において、面単位で不定回数の値加算をするためのユーティリティ関数}
\usage{
my.vector.access(v, a, func = sum, zero = 0)
}
\description{
Utility 関数
特に疎行列において、面単位で不定回数の値加算をするためのユーティリティ関数
}

