% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.whole.vol.R
\name{my.whole.vol}
\alias{my.whole.vol}
\title{Featuring 3D+1T array data
The functions are to quantitate various features of 3D+1T data}
\usage{
my.whole.vol(v)
}
\description{
Featuring 3D+1T array data
The functions are to quantitate various features of 3D+1T data
}
\examples{
library(igraph)
df <- 4
n <- 20
data. <- array(0,rep(n,df))
addr <- t(which(data.>=0,arr.ind=TRUE))
n.sp <- 20
for(i in 1:n.sp){
	tmp <- runif(df) * n
	tmp2 <- runif(1) * n/5
	data.[apply((addr - tmp)^2,2,sum) < tmp2] <- 1
}
clu <- my.labeling(data.)
hist(clu[[3]],ylim=c(0,100))
cluster.n <- clu[[1]][[3]]
vol <- vol.slice <- list()
ctr <- ctr.slice <- list()
tri.mesh.slice <- list()
dm <- dim(data.)
for(i in 1:cluster.n){
	tmp.arr <- array(as.numeric(clu[[3]]==i),dm)
	vol[[i]] <- my.whole.vol(tmp.arr)
	vol.slice[[i]] <- my.slice.vol(tmp.arr,df)
	ctr[[i]] <- my.whole.center(tmp.arr)
	ctr.slice[[i]] <- my.slice.center(tmp.arr,df)
	tri.mesh.slice[[i]] <- my.slice.tri.mesh(tmp.arr.df)
}
}

