% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.DECconformal.R
\name{plot.sp.conformal}
\alias{plot.sp.conformal}
\title{３列の実行列(３次元座標の行列)xyzと、３行の頂点ID行列faces.vとを用いて、三角形メッシュを塗り描く関数
ただし、この関数は、元がsp.triは、元の三角形メッシュがmy.sphere.tri.mesh()によって作られていることを前提としている。そのような場合に、sp.triという引数によって、隣接緯線間ごとに色を塗り分けることで、元の正球からの変形の様子を色表現することができる
また、各三角形の色はrho.fなる実ベクトルによってコントラストをつけることができる
col1は、元の正球からの変化を縞模様化するための長さ２の整数ベクトルである。その縞模様を使いたくないとき(メッシュがmy.sphere.tri.mesh()由来ではないときを含む)は、この長さ２のベクトルの２つの値を同じにすればよい)}
\usage{
\method{plot}{sp.conformal}(xyz, faces.v, sp.tri, rho.f, col1 = c(4, 5))
}
\description{
３列の実行列(３次元座標の行列)xyzと、３行の頂点ID行列faces.vとを用いて、三角形メッシュを塗り描く関数
ただし、この関数は、元がsp.triは、元の三角形メッシュがmy.sphere.tri.mesh()によって作られていることを前提としている。そのような場合に、sp.triという引数によって、隣接緯線間ごとに色を塗り分けることで、元の正球からの変形の様子を色表現することができる
また、各三角形の色はrho.fなる実ベクトルによってコントラストをつけることができる
col1は、元の正球からの変化を縞模様化するための長さ２の整数ベクトルである。その縞模様を使いたくないとき(メッシュがmy.sphere.tri.mesh()由来ではないときを含む)は、この長さ２のベクトルの２つの値を同じにすればよい)
}

