% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my.catmull.clark.tri.R
\name{my.catmull.clark.tri}
\alias{my.catmull.clark.tri}
\title{3D triangle mesh smoothing}
\usage{
my.catmull.clark.tri(x.vet)
}
\arguments{
\item{x.vet}{is a list. The first element is 3-column matrix of coordinates of vertices and the second element is triangular-mesh information object (output of my.tri.vet() ).}
}
\description{
These functions make triangle mesh smooth with Catmull-Clark subdivision
}
\examples{
library(rgl)
library(misc3d)
v <- matrix(c(1,2,3,2,3,1,3,3,3),byrow=TRUE,ncol=3)
u <- v[c(3,2,1,3,2,2,2,3),]
my.tri.v.id(v,u)
k <- 4
out <- my.3d.obj(3,k)
obj <- out$xyz[which(out$xyz.val==1),]
plot3d(obj)
xyz.arr <- array(out$xyz.val,c(2^k,2^k,2^k))
tris <- my.peri.tri(xyz.arr)
drawScene.rgl(tris)
sss <- my.tri.vet(ttt[[2]])
x.vet <- list(ttt[[1]],sss)
new.vet <- my.catmull.clark.tri(x.vet)
plot3d(new.vet[[1]][new.vet[[2]]$v.of.t,])
}
\keyword{mesh}
\keyword{shape}
\keyword{smoothing}
\keyword{triangulation}

