% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalk3d.R
\name{my.random.walk}
\alias{my.random.walk}
\title{3D+time movement}
\usage{
my.random.walk(Nobs, Nt, d)
}
\description{
3D + time movement
}
\examples{
library(rgl)
Nobs <- 5;Nt <- 100; d <- 3
X = my.random.walk(Nobs,Nt,d)
X.k <- my.k.diff(X)
X.d <- my.array.diff(X)
}
\keyword{movement}

