% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.spin.transformation.R
\name{my.spin.transformation}
\alias{my.spin.transformation}
\title{Spin transformation}
\usage{
my.spin.transformation(n = 5, k = runif(1), n.mesh = 32, n.step = 10,
  scale.shift = 0.1, scale.rotation1 = 0.1, scale.rotation2 = 0.1,
  scale.shear = 0.1, file.name = "hoge")
}
\arguments{
\item{n}{depth of sphere harmonic functions}

\item{k}{simple sphere radius}

\item{n.mesh}{fineness of mesh}

\item{n.step}{number of time steps}

\item{scale.shift}{scale of shift}

\item{scale.rotation1}{scale of rotation 1}

\item{scale.rotation2}{scale of rotation 2}

\item{scale.shear}{scale of shear}

\item{file.name}{output file name}
}
\description{
Spin transformation
}
\examples{
library(rgl)
library(RFOC)
library(onion)
out <- my.sphere.field()
# ”¼Œa‚ðF‚Å
col1 <- col2 <- rep(0,length(out$v[,1]))
col1[which(out$v[,1]>=0)] <- out$v[,1][which(out$v[,1]>=0)]
col2[which(out$v[,1]<0)] <- -out$v[,1][which(out$v[,1]<0)]
col1 <- col1/max(col1)
col2 <- col2/max(col2)
plot3d(out$mesh$x,col=rgb(1-col1,1-col2,0),size=10)
}

